api/requirements/base.txt
Django
djangorestframework
django-extensions
psycopg2
django-cors-headers
django-dotenv
django-filter
django_crontab
django-redis
drf-extensions
